+++ nextgen/action-graph/src/action_graph_error.rs
@@ -4,6 +4,7 @@ use thiserror::Error;
 
 #[derive(Error, Debug, Diagnostic)]
 pub enum ActionGraphError {
+    #[diagnostic(code(action_graph::cycle_detected))]
     #[error("A dependency cycle has been detected for {}.", .0.style(Style::Label))]
     CycleDetected(String),
 }